Initial Commit
Implemented URL extraction, domain parsing and checking against
a stoplist. The current top_domains list is drawn from an old
dump of Alexa rankings, so should be updated.

The code relies on my urlmorph's expander transform, which I've
not included because I'm going to figure out how to package
python libraries properly and then just make it a dependency.

Only the count of mentions is used so far in ranking the domains.